+++ quad_SL/c_code_SNN/functional.h
@@ -4,7 +4,7 @@
 void print_array_1d(int const size, float const x[size]);
 
 // Print 1D array of floats (as integers)
-void print_array_1d_bool(int const size, float const x[size]);
+void print_array_1d_bool(int const size, int const x[size]);
 
 // Print 2D array of floats (as floats)
 void print_array_2d(int const rows, int const cols, float const x[rows][cols]);
